+++ test/integration/include/find.test.js
@@ -170,6 +170,30 @@ describe(Support.getTestDialectTeaser('Include'), function() {
         });
     });
 
+    it('should support a belongsTo with the targetKey option', function() {
+      var User = this.sequelize.define('User', { username: DataTypes.STRING })
+        , Task = this.sequelize.define('Task', { title: DataTypes.STRING });
+      User.removeAttribute('id');
+      Task.belongsTo(User, { foreignKey: 'user_name', targetKey: 'username'});
+
+      return this.sequelize.sync({ force: true }).then(function() {
+        return User.create({ username: 'bob' }).then(function(newUser) {
+          return Task.create({ title: 'some task' }).then(function(newTask) {
+            return newTask.setUser(newUser).then(function() {
+              return Task.find({
+                title: 'some task',
+                include: [ { model: User } ]
+              })
+                .then(function (foundTask) {
+                  expect(foundTask).to.be.ok;
+                  expect(foundTask.User.username).to.equal('bob');
+                });
+            });
+          });
+        });
+      });
+    });
+
     it('should support many levels of belongsTo (with a lower level having a where)', function() {
       var A = this.sequelize.define('a', {})
         , B = this.sequelize.define('b', {})